[IMP] web: split column_width_hook into dedicated module
The change resolves the static reference issue where the computeWidths function
could not be patched. This limitation occurs because the `column_width_hook`
declares and invokes computeWidths directly within its body (specifically
in forceColumnWidths), creating a static closure at import time. As a result,
dynamic overrides were prevented, reducing flexibility for extending or
customizing behavior.

To address this, hardcoded constants, date width computation, and the computeWidths
logic were extracted from column_width_hook into a new column_widths module.
This modularization enables proper patching and extension while maintaining
compatibility with dependent modules.
